+++ examples/pyplot/histo_pca.py
@@ -1,6 +1,6 @@
 """Histogram along a PCA axis"""
 import numpy as np
-from vedo import Points, pcaEllipsoid, Arrow2D, Goniometer
+from vedo import Points, pcaEllipse, Arrow2D, Goniometer
 from vedo.pyplot import Figure, histogram
 
 data = np.random.randn(1000, 3)
@@ -10,7 +10,8 @@ pts.scale([2,1,0.01]).rotateZ(45).shift(5,1)  # rotate and shift!
 
 # Recover the rotation pretending we only know the points
 # Fit a 1-sigma ellipse to the points
-elli = pcaEllipsoid(pts)
+elli = pcaEllipse(pts)
+
 ec, e1, e2 = elli.center, elli.axis1, elli.axis2
 arrow1 = Arrow2D(ec, ec - 3*e1)
 arrow2 = Arrow2D(ec, ec + 3*e2)
@@ -29,5 +30,5 @@ histo.rotateZ(90 + angle).pos(ec - 6*e1)
 gon = Goniometer(ec-5.5*e1, ec, [ec[0]-5.5*e1[0], ec[1],0]).z(0.2)
 
 fig = Figure([0,14], [-4,9], aspect="equal", title=__doc__)
-fig += [pts, elli, arrow1, arrow2, gon, histo]
+fig += [pts, elli.z(-0.1), arrow1, arrow2, gon, histo]
 fig.show(zoom='tight').close()